+++ CHANGELOG.md
@@ -67,7 +67,9 @@
 
 - Refactored `BacktraceStorage` to simplify the code.
 - Improved `LoggerBase` memory layout by reorganizing class members.
-- Fix multiple definitions of `on_alarm` in `SignalHandler.h`
+- Fixed multiple definitions of `on_alarm` in `SignalHandler.h`
+- Fixed a bug in the backend thread where flush() and run_periodic_tasks() were skipped for certain sinks. All unique
+  sinks are now correctly processed.
 
 ## v4.2.1