+++ CHANGELOG.md
@@ -68,8 +68,8 @@
 - Refactored `BacktraceStorage` to simplify the code.
 - Improved `LoggerBase` memory layout by reorganizing class members.
 - Fixed multiple definitions of `on_alarm` in `SignalHandler.h`
-- Fixed a bug in the backend thread where flush() and run_periodic_tasks() were skipped for certain sinks. All unique
-  sinks are now correctly processed.
+- Fixed a bug in the backend thread where `flush()` and `run_periodic_tasks()` were skipped for certain sinks. All sinks
+  are now correctly processed.
 
 ## v4.2.1